refactor storage tool
